makes changes as requested
