--- templates/views/browse/browse.html
@@ -1,79 +1,0 @@
-<!-- Include Top Navbar -->
-{{>top-nav}}
-
-<!-- Main Content -->
-<div class="container">
-    <div class="card card--header daintree--blue">
-      {{#if view.challenges}}
-      <a class="title--link" href="/challenges/tags"><h1 class="title">
-        {{#if filters}}{{filters.name}}{{else}}All tags{{/if}}
-      {{else if view.orgs}}
-      <a class="title--link" href="/orgs/tags"><h1 class="title">
-        {{#if filters}}{{filters.name}}{{else}}All tags{{/if}}
-      {{/if}}
-      </h1>
-      <div class="icon icon--dropdown"></div>
-      </a>
-    </div>
-
-    <!-- TODO MAKE THIS DIV A PARTIAL -->
-    <div class="card card--header">
-      <ul class="navbar__links">
-        {{#if view.challenges}}
-          <li class="navbar__item togglebar__item">
-            {{#if filters}}
-            <a class="navbar__link togglebar__link" href="/orgs?tags={{filters.id}}">
-            {{else}}
-            <a class="navbar__link togglebar__link" href="/orgs">
-            {{/if}}
-              Organisations
-          </a></li>
-
-          <li class="navbar__item togglebar__item--selected">
-            {{#if filters}}
-            <a class="navbar__link togglebar__link togglebar__link--selected" href="/challenges?tags={{filters.id}}">
-            {{else}}
-            <a class="navbar__link togglebar__link togglebar__link--selected" href="/challenges">
-            {{/if}}
-              Challenges
-          </a></li>
-
-        {{else if view.orgs}}
-          <li class="navbar__item togglebar__item--selected">
-            {{#if filters}}
-            <a class="navbar__link togglebar__link togglebar__link--selected" href="/orgs?tags={{filters.id}}">
-              {{else}}
-              <a class="navbar__link togglebar__link togglebar__link--selected" href="/orgs">
-                {{/if}}
-              Organisations
-          </a></li>
-
-          <li class="navbar__item togglebar__item">
-            {{#if filters}}
-            <a class="navbar__link togglebar__link" href="/challenges?tags={{filters.id}}">
-            {{else}}
-            <a class="navbar__link togglebar__link" href="/challenges">
-            {{/if}}
-            Challenges
-            </a></li>
-        {{/if}}
-      </ul>
-    </div>
-
-      {{#if data}}
-      {{#if view.challenges}}
-        {{#each data}}
-          {{>challenge}}
-        {{/each}}
-      {{else}}
-        {{#each data}}
-          {{>organisation}}
-        {{/each}}
-      {{/if}}
-      {{else}}
-      <h3 class="title title-blue">Nothing meets your search criteria.</h3>
-      {{/if}}
-</div>
-
-<!-- Include Navbar -->
-{{>navbar}}
